Moved models to its own namespace
