+++ tests/Fixture/Model/Employee.php
@@ -8,7 +8,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Tests\TSantos\Serializer\Fixture;
+namespace Tests\TSantos\Serializer\Fixture\Model;
 
 /**
  * Class Employee
