+++ tests/Metadata/Driver/XmlDriverTest.php
@@ -25,7 +25,7 @@ class XmlDriverTest extends AbstractDriverTest
     public function createDriver(): DriverInterface
     {
         return new XmlDriver(new FileLocator([
-            'Tests\TSantos\Serializer\Fixture' => __DIR__ . '/../../Resources/mapping']
+            'Tests\TSantos\Serializer\Fixture\Model' => __DIR__ . '/../../Resources/mapping']
         ), new TypeGuesser());
     }
 }
