+++ tests/Metadata/Driver/YamlDriverTest.php
@@ -33,7 +33,7 @@ class YamlDriverTest extends AbstractDriverTest
     public function createDriver(): DriverInterface
     {
         return new YamlDriver(new FileLocator([
-            'Tests\TSantos\Serializer\Fixture' => __DIR__ . '/../../Resources/mapping']
+            'Tests\TSantos\Serializer\Fixture\Model' => __DIR__ . '/../../Resources/mapping']
         ), new TypeGuesser());
     }
 }
