+++ tests/Resources/mapping/Person.xml
@@ -9,14 +9,14 @@
   -->
 
 <serializer>
-    <class name="Tests\TSantos\Serializer\Fixture\Person" base-class="Tests\TSantos\Serializer\AbstractSerializerClass">
+    <class name="Tests\TSantos\Serializer\Fixture\Model\Person" base-class="Tests\TSantos\Serializer\AbstractSerializerClass">
         <property name="id" type="integer" />
         <property name="name" type="string" expose-as="nome" groups="api" />
         <property name="lastName" type="string" />
         <property name="married" type="boolean" getter="isMarried" expose-as="is_married" />
         <property name="birthday" type="DateTime" modifier="format('d/m/Y')" />
         <property name="address" read-only="true" />
-        <property name="father" type="Tests\TSantos\Serializer\Fixture\Person" />
+        <property name="father" type="Tests\TSantos\Serializer\Fixture\Model\Person" />
         <virtual_property name="getFullName" type="string" expose-as="full_name" >
             <groups>
                 <value>api</value>
