+++ tests/Serialization/NormalizationTest.php
@@ -10,8 +10,8 @@
 
 namespace Tests\TSantos\Serializer\Serialization;
 
-use Tests\TSantos\Serializer\Fixture\Book;
-use Tests\TSantos\Serializer\Fixture\Person;
+use Tests\TSantos\Serializer\Fixture\Model\Book;
+use Tests\TSantos\Serializer\Fixture\Model\Person;
 use Tests\TSantos\Serializer\SerializerTestCase;
 
 /**
